Fixing protocol number types
Copying and pasting code blocks was a bad idea

+++ src/internet/model/icmpv4-l4-protocol.h
@@ -51,9 +51,9 @@ public:
    */
   static TypeId GetTypeId (void);
 #ifndef _MSC_VER
-        static const uint16_t PROT_NUMBER;//!< ICMP protocol number (0x1)
+        static const uint8_t PROT_NUMBER;//!< ICMP protocol number (0x1)
 #else
-        constexpr static const uint16_t PROT_NUMBER = 0x1;
+        constexpr static const uint8_t PROT_NUMBER = 0x1;
 #endif
   Icmpv4L4Protocol ();
   virtual ~Icmpv4L4Protocol ();
