+++ src/internet/model/udp-l4-protocol.h
@@ -66,9 +66,9 @@ public:
    */
   static TypeId GetTypeId (void);
 #ifndef _MSC_VER
-        static const uint16_t PROT_NUMBER;//!< protocol number (0x11)
+        static const uint8_t PROT_NUMBER;//!< protocol number (0x11)
 #else
-        constexpr static const uint16_t PROT_NUMBER = 0x11;
+        constexpr static const uint8_t PROT_NUMBER = 0x11;
 #endif
   UdpL4Protocol ();
   virtual ~UdpL4Protocol ();
